+++ core/scripts/keystone/src/03_gen_crib_cluster_overrides_cmd.go
@@ -51,11 +51,7 @@ func generateCribConfig(pubKeysPath string, chainID *int64, templatesDir string,
 
 	lines, err := readLines(filepath.Join(templatesDir, cribOverrideTemplate))
 	helpers.PanicErr(err)
-	lines = replaceCribPlaceholders(
-		lines,
-		forwarderAddress,
-		nodeAddresses,
-	)
+	lines = replaceCribPlaceholders(lines, forwarderAddress, nodeAddresses)
 	return lines
 }
