+++ core/scripts/keystone/src/99_errors.go
@@ -1,0 +1,8 @@
+package src
+
+// PanicErr panics if error the given error is non-nil.
+func PanicErr(err error) {
+	if err != nil {
+		panic(err)
+	}
+}
